+++ docs/web_modules/LayoutContainer/index.js
@@ -4,9 +4,11 @@ import Helmet from "react-helmet"
 import Header from "../Header"
 import Footer from "../Footer"
 import GoogleAnalyticsTracker from "../GoogleAnalyticsTracker"
+import AppCacheBanner from "../AppCacheBanner"
 
+import "./index.global.css"
+import "./hightlightjs.global.css"
 import styles from "./index.css"
-import "./hightlightjs.css"
 
 export default class Layout extends Component {
 
@@ -42,7 +44,17 @@ export default class Layout extends Component {
             { src: "https://cdn.polyfill.io/v2/polyfill.min.js" },
           ] }
         />
+
+        { /* meta viewport safari/chrome/edge */ }
+        <Helmet
+          meta={ [ {
+            name: "viewport", content: "width=device-width, initial-scale=1",
+          } ] }
+        />
+        <style>{ "@-ms-viewport { width: device-width; }" }</style>
+
         <div className={ styles.layout }>
+          <AppCacheBanner />
           <Header />
           <div className={ styles.content }>
             { this.props.children }