+++ docs/web_modules/layouts/PageLoading/index.js
@@ -1,4 +1,5 @@
 import React, { Component } from "react"
+import Helmet from "react-helmet"
 import TopBarProgressIndicator from "react-topbar-progress-indicator"
 
 import Banner from "../../Banner"
@@ -18,6 +19,9 @@ export default class PageLoading extends Component {
   render() {
     return (
       <div>
+        <Helmet
+          title={ "Loading..." }
+        />
         <TopBarProgressIndicator />
         <Banner small>
           <h1>{ " " }</h1>